airflow/requirements.txt
# airflow/requirements.txt
# Airflow 컨테이너에 추가 설치할 패키지

# 문서 파싱
pdfminer.six>=20221105
pypdf>=4.1.0
python-docx>=1.1.0
beautifulsoup4>=4.12.0

# HTTP 크롤링
httpx>=0.27.0

# 데이터베이스
pymilvus>=2.4.0
psycopg2-binary>=2.9.9

# AI / 임베딩
numpy==1.26.4
torch==2.2.2+cpu --index-url https://download.pytorch.org/whl/cpu
sentence-transformers==2.7.0
transformers==4.40.2
tokenizers==0.19.1

# PII 마스킹
presidio-analyzer>=2.2.0
presidio-anonymizer>=2.2.0
